Refactor models and migrations for Places app
- Consolidated `spatial.py` and `sites.py` into a single `models.py` file.
- Updated `Department`, `Municipality`, `Locality`, and `Neighborhood` models:
     - Removed `unique=True` from `name` fields.
     - Added unique constraints for `name` combined with parent relationships.
- Updated migrations to reflect model changes.
- Removed unused `__init__.py` and `sites.py` files from `models` directory.
