--- backend/apps/places/models/__init__.py
@@ -1,12 +1,0 @@
-from .spatial import Country, Department, Municipality, Locality, Neighborhood
-from .sites import Site
-
-
-__all__ = [
-    "Country",
-    "Department",
-    "Municipality",
-    "Locality",
-    "Neighborhood",
-    "Site",
-]
